Removed protected flag from password status.
Not sure if this was intentional or not, but the "pwned"/"secure" status text was being stored in a protected string, which defaults to showing Asterisks when the field is displayed with other plugins, such as KPEnhancedEntryView. Since the text is intended to be visible to the user and not designed to hold sensitive information the flag can be safely removed, allowing greater compatibility with other plugins/display methods.