Set Status field as unprotected
When setting the field/column HIBP is storing the pwned status as a protected string. This is a bit counter productive as it hides the breach status (which in and of itself does not contain the password) behind asterisks ******. What good is it to check the breach status if you can't see the result without extra effort?

This mainly affects views that don't use the "Column". Such as the text view (and plugins that make use of the text view) and plugins such as EnhancedKeepassEntryView that provide alternate views/layouts. The column view defaults to unmasking the field.